Fuck me, note the sd card does not work

+++ weather_node/main/pcnt_sensor.c
@@ -106,7 +106,7 @@ esp_err_t get_wind_read(sensor_payload_t *payload) {
     int wind_cnt = 0;
     int64_t cur_time = esp_timer_get_time();
     int64_t duration = cur_time - wind_last_time;
-    float mms = 0;
+    double mms = 0;
 
     if (duration <= 0)
         return ESP_FAIL;
@@ -114,7 +114,7 @@ esp_err_t get_wind_read(sensor_payload_t *payload) {
     ESP_RETURN_ON_ERROR(pcnt_unit_get_count(wind_pcnt_unit, &wind_cnt), TAG, "WIND> Failed to read count");
     pcnt_unit_clear_count(wind_pcnt_unit);
 
-    mms = PI * WIND_CUP_DIAMETER * CALIBRATION_FACTOR * wind_cnt * 1000000000ULL / duration;
+    mms = PI * WIND_CUP_DIAMETER * CALIBRATION_FACTOR * wind_cnt * (double)(1000000000.0 / duration);
     payload->wind_speed = (uint16_t)(mms);
     // ESP_LOGI(TAG, "pulses = %ld, mms = %f, shared = %d", wind_cnt, (double)mms, wind_shared_speed);
